Adjustment in entities; Repositories rebuilt; Adjustment in Compose.yaml; Application.yaml was add

+++ src/main/java/com/example/simpletask/Models/Entity/CategoryProduct_Entity.java
@@ -17,7 +17,7 @@ public class CategoryProduct_Entity {
     @Id
     @Column(name = "CategoryId",nullable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long CategoryId;
+    private Integer CategoryId;
 
     @Column(name = "Category_Name",nullable = false)
     private String Category_Name;
